requirements.txt
pandas==1.3.3
psycopg2-binary==2.9.1
flask==2.0.1
sqlalchemy==1.3.20
